Add files via upload
The algorithms used in this prediction are the regression algorithms, such as XGBoost regressor, Multiple Linear Regression, and RandomForest regressor. Usually, the dataset is full of noise and redundancy, for that some preprocessing techniques have been applied for the preparation of data. The models embedded with algorithms have gone through some  hyperparameter tuning so as to express their best performance with the tuned and appropriate parameters.